tests/wpt/tests/html/webappapis/system-state-and-capabilities/the-navigator-object/navigatorcookies-cookieenabled-false-manual.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>navigator.cookieEnabled false</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="help" href="https://html.spec.whatwg.org/#dom-navigator-cookieenabled">
<meta name="flags" content="interact">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<h2>Preconditions</h2>
<p>Disable cookies in browser settings.</p>

<script>
  test(() => {
    assert_false(navigator.cookieEnabled);
  }, "navigator.cookieEnabled is false when cookies are disabled");
</script>
